fix: removed content field of Article entity
In Article entity, content field was holding a direct parsed markdown from
markdown field in html format. Meaning that if article change the way it
parses markdown, the blog admin would have to resave every article in
order for changes to apply. Also it allows to parse code block for
syntax highlighting using the markdown parser, making it easier.